creating a link_functions.jl
